Fix Ctrl+Backspace triggering Ctrl+Del hotkeys

+++ StableDiffusionGui/Ui/MainUiHotkeys.cs
@@ -1,4 +1,6 @@
 ﻿using StableDiffusionGui.Forms;
+using StableDiffusionGui.Main;
+using StableDiffusionGui.MiscUtils;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -6,6 +8,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using System.Windows.Input;
 
 namespace StableDiffusionGui.Ui
 {
@@ -22,10 +25,10 @@ namespace StableDiffusionGui.Ui
             if (keys == (Keys.Control | Keys.G)) // Hotkey: Generate/Cancel
                 Program.MainForm.RunBtn.PerformClick();
 
-            if (keys == (Keys.Control | Keys.Delete)) // Hotkey: Delete generated image
+            if (keys == (Keys.Control | Keys.Delete) && !InputUtils.IsKeyPressed(Key.Back)) // Hotkey: Delete generated image
                 ImagePreview.DeleteCurrent();
 
-            if (keys == (Keys.Control | Keys.Alt | Keys.Delete)) // Hotkey: Delete all generated images
+            if (keys == (Keys.Control | Keys.Alt | Keys.Delete) && !InputUtils.IsKeyPressed(Key.Back)) // Hotkey: Delete all generated images
                 ImagePreview.DeleteAll();
 
             if (keys == (Keys.Control | Keys.Add) || keys == (Keys.Control | Keys.Oemplus)) // Hotkey: Toggle prompt field size
